renamed `cfi.rs` in colorants `cfi_data.rs`, to distinguish from the CFI calculation module `cfi.rs`
